+++ src/util/ssa_expr.h
@@ -80,19 +80,19 @@ public:
 
   void set_level_0(unsigned i)
   {
-    set(ID_L0, i);
+    set(ID_L0, static_cast<long long>(i));
     update_identifier();
   }
 
   void set_level_1(unsigned i)
   {
-    set(ID_L1, i);
+    set(ID_L1, static_cast<long long>(i));
     update_identifier();
   }
 
   void set_level_2(unsigned i)
   {
-    set(ID_L2, i);
+    set(ID_L2, static_cast<long long>(i));
     update_identifier();
   }
